Move form engine auto configurations into dedicated packages
