+++ modules/flowable-spring-boot/flowable-spring-boot-starters/flowable-spring-boot-starter-basic/src/test/java/org/flowable/test/spring/boot/FormEngineAutoConfigurationTest.java
@@ -22,9 +22,9 @@ import org.flowable.form.api.FormService;
 import org.flowable.form.engine.FormEngine;
 import org.flowable.form.engine.FormEngineConfiguration;
 import org.flowable.spring.boot.FlowableTransactionAutoConfiguration;
-import org.flowable.spring.boot.FormEngineAutoConfiguration;
-import org.flowable.spring.boot.FormEngineServicesAutoConfiguration;
 import org.flowable.spring.boot.ProcessEngineAutoConfiguration;
+import org.flowable.spring.boot.form.FormEngineAutoConfiguration;
+import org.flowable.spring.boot.form.FormEngineServicesAutoConfiguration;
 import org.junit.Test;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceTransactionManagerAutoConfiguration;
